7.31

+++ QiangGouHui/QGHFirstPageGoodsModel.h
@@ -14,7 +14,7 @@
 @property (nonatomic, copy) NSString *area;
 @property (nonatomic, assign) NSInteger create_time;
 @property (nonatomic, assign) float discount_price;
-@property (nonatomic, assign) NSInteger end_time;
+@property (nonatomic, assign) double end_time;
 @property (nonatomic, copy) NSString *goodsId;
 @property (nonatomic, assign) float original_price;
 @property (nonatomic, assign) NSInteger production_time;
